backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4
numpy==1.26.2
yfinance==0.2.65
